requirements.txt
feedparser>=6.0.0
requests>=2.31.0
google-generativeai>=0.8.0
google-genai>=1.0.0
tenacity>=8.2.0
Pillow>=10.0.0
